requirements.txt
langchain==0.0.177
gpt4all==0.2.3
chromadb==0.3.23
llama-cpp-python==0.1.50
urllib3==2.0.2
python-dotenv==1.0.0
tqdm==4.65.0
wikipedia-api==0.5.8

